Crate eventstore

Re-exports

pub use types::*;

Modules

commands
Commands this client supports.

types
Common types used across the library.

Structs

Connection
Represents a connection to a single node. Client maintains a full duplex connection to the EventStore server. An EventStore connection operates quite differently than say a SQL connection. Normally when you use an EventStore connection you want to keep the connection open for a much longer of time than when you use a SQL connection.

ConnectionBuilder
Helps constructing a connection to the server.