Struct eventstore::commands::ReadStreamEvents

pub struct ReadStreamEvents<'a> { /* fields omitted */ }

A command that reads several events from a stream. It can read events forward or backward.

Methods

impl<'a> ReadStreamEvents<'a>

pub fn forward(self) -> ReadStreamEvents<'a>

Asks the command to read forward (toward the end of the stream). That's the default behavior.

pub fn backward(self) -> ReadStreamEvents<'a>

Asks the command to read backward (toward the begining of the stream).

pub fn credentials(self, value: Credentials) -> ReadStreamEvents<'a>

Performs the command with the given credentials.

pub fn max_count(self, max_count: i32) -> ReadStreamEvents<'a>

Max batch size.

pub fn start_from(self, start: i64) -> ReadStreamEvents<'a>

Starts the read at the given event number. By default, it starts at 0.

pub fn start_from_beginning(self) -> ReadStreamEvents<'a>

Starts the read from the beginning of the stream. It also set the read direction to Forward.

pub fn start_from_end_of_stream(self) -> ReadStreamEvents<'a>

Starts the read from the end of the stream. It also set the read direction to Backward.

pub fn require_master(self, require_master: bool) -> ReadStreamEvents<'a>

Asks the server receiving the command to be the master of the cluster in order to perform the write. Default: false.

pub fn resolve_link_tos(self, tos: LinkTos) -> ReadStreamEvents<'a>

When using projections, you can have links placed into another stream. If you set true, the server will resolve those links and will return the event that the link points to. Default: NoResolution.

pub fn execute(
self
) -> impl Future<Item = ReadStreamStatus<StreamSlice>, Error = OperationError>

Sends asynchronously the read command to the server.

pub fn iterate_over_batch(
self
) -> impl Stream<Item = Vec<ResolvedEvent>, Error = OperationError> + 'a

Returns a Stream that consumes a stream entirely. For example, if the direction is Forward, it ends when the last stream event is reached. However, if the direction is Backward, the iterator ends when the first event is reached. All the configuration is pass to the iterator (link resolution, require master, starting point, batch size, …etc). Each element corresponds to a page with a length <= max_count.

pub fn iterate_over(
self
) -> impl Stream<Item = ResolvedEvent, Error = OperationError> + 'a

Returns a Stream that consumes a stream entirely. For example, if the direction is Forward, it ends when the last stream event is reached. However, if the direction is Backward, the iterator ends when the first event is reached. All the configuration is pass to the iterator (link resolution, require master, starting point, batch size, …etc).