Struct eventstore::types::GossipSeedClusterSettings

pub struct GossipSeedClusterSettings { /* fields omitted */ }

Contains settings related to a cluster of fixed nodes.

Methods

impl GossipSeedClusterSettings

pub fn new(seeds: Vec1<GossipSeed>) -> GossipSeedClusterSettings

Creates a GossipSeedClusterSettings from a non-empty list of gossip seeds.

pub fn set_gossip_timeout(
self,
gossip_timeout: Duration
) -> GossipSeedClusterSettings

Maximum duration a node should take when requested a gossip request.

pub fn set_max_discover_attempts(
self,
max_attempt: usize
) -> GossipSeedClusterSettings

Maximum number of retries during a discovery process. Discovery process is when the client tries to figure out the best node to connect to.